moved getNodeAccountMap from constants to helpers
